+++ api/src/main/java/it/cs/unicam/MunicipalDigitalization/api/util/DesignPattern/Builder/AuthorizedItineraryBuilder.java
@@ -1,4 +1,4 @@
-package it.cs.unicam.MunicipalDigitalization.api.util.DesignPattern;
+package it.cs.unicam.MunicipalDigitalization.api.util.DesignPattern.Builder;
 
 import it.cs.unicam.MunicipalDigitalization.api.model.Municipality;
 import it.cs.unicam.MunicipalDigitalization.api.model.actors.AbstractAuthenticatedUser;
@@ -8,11 +8,13 @@ import it.cs.unicam.MunicipalDigitalization.api.util.Coordinate;
 import it.cs.unicam.MunicipalDigitalization.api.util.ElementStatus;
 import it.cs.unicam.MunicipalDigitalization.api.util.MatchingAlgorithms;
 import it.cs.unicam.MunicipalDigitalization.api.util.UserRole;
+import org.springframework.stereotype.Component;
 
 import java.util.List;
 
 import static it.cs.unicam.MunicipalDigitalization.api.util.MatchingAlgorithms.containsSpecialCharacters;
 
+@Component
 public class AuthorizedItineraryBuilder implements ItineraryBuilder{
     private List<AbstractPOI> poiList;
     private String name;
@@ -53,7 +55,7 @@ public class AuthorizedItineraryBuilder implements ItineraryBuilder{
 
     @Override
     public void setItineraryAuthor(AbstractAuthenticatedUser author) {
-        if(author.getRole() == UserRole.AUTHORIZED_CONTRIBUTOR || author.getRole() == UserRole.CURATOR) this.author = author;
+        if(author.getRole().contains(UserRole.AUTHORIZED_CONTRIBUTOR )|| author.getRole().contains(UserRole.CURATOR)) this.author = author;
         else throw new IllegalArgumentException("The author is not a contributor");
     }
