+++ api/src/main/java/it/cs/unicam/MunicipalDigitalization/api/util/DesignPattern/Builder/ContributionContestBuilder.java
@@ -1,0 +1,4 @@
+package it.cs.unicam.MunicipalDigitalization.api.util.DesignPattern.Builder;
+
+public interface ContributionContestBuilder {
+}
