--- api/src/main/java/it/cs/unicam/MunicipalDigitalization/api/util/DesignPattern/ContributionContestBuilder.java
@@ -1,4 +1,0 @@
-package it.cs.unicam.MunicipalDigitalization.api.util.DesignPattern;
-
-public interface ContributionContestBuilder {
-}
